+++ config/api/resource-type/parameters.php
@@ -5,18 +5,18 @@ declare(strict_types=1);
 return [
     'collection' => [
         'include_resources' => [
-            'field' => 'include_resources',
-            'title' => 'resource-type/parameters.title-include_resources',
-            'description' => 'resource-type/parameters.description-include_resources',
+            'field' => 'include-resources',
+            'title' => 'resource-type/parameters.title-include-resources',
+            'description' => 'resource-type/parameters.description-include-resources',
             'type' => 'boolean',
             'required' => false
         ]
     ],
     'item' => [
         'include_resources' => [
-            'field' => 'include_resources',
-            'title' => 'resource-type/parameters.title-include_resources',
-            'description' => 'resource-type/parameters.description-include_resources',
+            'field' => 'include-resources',
+            'title' => 'resource-type/parameters.title-include-resources',
+            'description' => 'resource-type/parameters.description-include-resources',
             'type' => 'boolean',
             'required' => false
         ]